noteworth_challenge_app/pyproject.toml
[tool.poetry]
name = "noteworth_challenge_app"
version = "0.1.0"
description = ""
authors = ["alexaled <alexaled@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.5"
requests = "^2.25.1"
djangorestframework = "^3.12.2"
celery = "^5.0.5"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
pytest-django = "^4.1.0"
responses = "^0.12.1"
pytest = "^6.2.2"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
